+++ core-services/auth-service/src/main/java/com/piggymetrics/auth/AuthApplication.java
@@ -1,10 +1,29 @@
 package com.piggymetrics.auth;
 
+import com.piggymetrics.auth.service.security.MongoUserDetailsService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.NoOpPasswordEncoder;
+import org.springframework.security.oauth2.config.annotation.configurers.ClientDetailsServiceConfigurer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.AuthorizationServerConfigurerAdapter;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableAuthorizationServer;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configurers.AuthorizationServerEndpointsConfigurer;
+import org.springframework.security.oauth2.config.annotation.web.configurers.AuthorizationServerSecurityConfigurer;
+import org.springframework.security.oauth2.provider.token.TokenStore;
+import org.springframework.security.oauth2.provider.token.store.InMemoryTokenStore;
 
 @SpringBootApplication
 @EnableResourceServer
@@ -14,5 +33,100 @@ public class AuthApplication {
 
   public static void main(String[] args) {
     SpringApplication.run(AuthApplication.class, args);
+  }
+
+  @Configuration
+  /**
+   * @EnableWebSecurity The configuration creates a Servlet Filter known as the springSecurityFilterChain which is responsible for all the security
+   * WebSecurityConfig only contains information about how to authenticate our users.
+   */
+  @EnableWebSecurity
+  /**
+   * WebSecurityConfigurerAdapter
+   * How does Spring Security know that we want to require all users to be authenticated? How does Spring Security know we want to support form based authentication
+   */
+  protected static class webSecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Autowired private MongoUserDetailsService userDetailsService;
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+      http.authorizeRequests().anyRequest().authenticated().and().csrf().disable();
+    }
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+      auth.userDetailsService(userDetailsService).passwordEncoder(new BCryptPasswordEncoder());
+    }
+    // for @Autowired private AuthenticationManager authenticationManager;
+    // Caused by: org.springframework.beans.factory.NoSuchBeanDefinitionException: No qualifying
+    // bean of type 'org.springframework.security.authentication.AuthenticationManager' available:
+    // expected at least 1 bean which qualifies as autowire candidate. Dependency annotations:
+    // {@org.springframework.beans.factory.annotation.Autowired(required=true)}
+    @Override
+    @Bean
+    public AuthenticationManager authenticationManagerBean() throws Exception {
+      return super.authenticationManagerBean();
+    }
+  }
+
+  @Configuration
+  @EnableAuthorizationServer
+  protected static class OAuth2AuthorizationConfig extends AuthorizationServerConfigurerAdapter {
+
+    private TokenStore tokenStore = new InMemoryTokenStore();
+
+    @Autowired private AuthenticationManager authenticationManager;
+
+    @Autowired private MongoUserDetailsService userDetailsService;
+
+    @Autowired private Environment env;
+
+    @Override
+    public void configure(ClientDetailsServiceConfigurer clients) throws Exception {
+      // TODO persist clients details
+      clients
+          .inMemory()
+          .withClient("browser")
+          .authorizedGrantTypes("refresh_token", "password")
+          .scopes("ui")
+          .and()
+          .withClient("account-service")
+          .secret(env.getProperty("ACCOUNT_SERVICE_PASSWORD"))
+          .authorizedGrantTypes("client_credentials", "refresh_token")
+          .scopes("server")
+          .and()
+          .withClient("statistics-service")
+          .secret(env.getProperty("STATISTICS_SERVICE_PASSWORD"))
+          .authorizedGrantTypes("client_credentials", "refresh_token")
+          .scopes("server")
+          .and()
+          .withClient("notification-service")
+          .secret(env.getProperty("NOTIFICATION_SERVICE_PASSWORD"))
+          .authorizedGrantTypes("client_credentials", "refresh_token")
+          .scopes("server");
+    }
+
+    @Override
+    public void configure(AuthorizationServerEndpointsConfigurer endpoints) {
+      endpoints
+          .tokenStore(tokenStore)
+          .authenticationManager(authenticationManager)
+          .userDetailsService(userDetailsService);
+    }
+
+    @Override
+    public void configure(AuthorizationServerSecurityConfigurer oauthServer) {
+      oauthServer.tokenKeyAccess("permitAll()").checkTokenAccess("isAuthenticated()");
+    }
+  }
+
+  // Troubleshooting
+  // https://docs.spring.io/spring-security/site/docs/current/reference/htmlsingle/#getting-started-experience
+  // java.lang.IllegalArgumentException: There is no PasswordEncoder mapped for the id "null"
+  @SuppressWarnings("deprecation")
+  @Bean
+  public static NoOpPasswordEncoder passwordEncoder() {
+    return (NoOpPasswordEncoder) NoOpPasswordEncoder.getInstance();
   }
 }
